Refactor Goby lib's exec implementation.
